app/requirements.txt
###### Requirements without Version Specifiers ######`
requests

###### Requirements with Version Specifiers ######`
google-cloud-secret-manager >= 2.2.0
google-cloud-logging >= 2.2.0